+++ design-system/docs/storybook/index.ts
@@ -1,0 +1,2 @@
+export * from '../node_modules/@storybook/react';
+export * from '../node_modules/@storybook/addon-actions';