+++ docs/storybook/index.ts
@@ -1,2 +1,0 @@
-export * from '@storybook/react';
-export * from '@storybook/addon-actions';