+++ keystatic/app/onboarding/setup.tsx
@@ -5,9 +5,9 @@ import { Box, Flex } from '@voussoir/layout';
 import { css } from '@voussoir/style';
 import { TextField } from '@voussoir/text-field';
 import { Heading, Text } from '@voussoir/typography';
-import { Config } from '../../src';
+import { GitHubConfig } from '../../src';
 
-export function KeystaticSetup(props: { config: Config }) {
+export function KeystaticSetup(props: { config: GitHubConfig }) {
   const [deployedURL, setDeployedURL] = useState('');
   const [organization, setOrganization] = useState('');
   return (
@@ -68,7 +68,7 @@ export function KeystaticSetup(props: { config: Config }) {
           name="manifest"
           className={css({ display: 'none' })}
           value={JSON.stringify({
-            name: `${props.config.repo.owner} Keystatic`,
+            name: `${props.config.storage.repo.owner} Keystatic`,
             url: deployedURL ? `${deployedURL}/keystatic` : 'http://localhost:3000/keystatic',
             public: false,
             redirect_url: 'http://localhost:3000/api/keystatic/github/created-app',