+++ keystatic/app/shell/data.tsx
@@ -1,8 +1,16 @@
 import { OperationData } from '@ts-gql/tag';
 import { gql } from '@ts-gql/tag/no-transform';
 import { useRouter } from 'next/router';
-import { Config } from '../../config';
-import { useMemo, useEffect, createContext, useContext, useCallback, ReactNode } from 'react';
+import { Config, GitHubConfig, LocalConfig } from '../../config';
+import {
+  useMemo,
+  useEffect,
+  createContext,
+  useContext,
+  useCallback,
+  ReactNode,
+  useState,
+} from 'react';
 import { useQuery } from 'urql';
 import { githubRequest } from '../../github-api';
 import {
@@ -11,20 +19,77 @@ import {
   getDataFileExtension,
   getSingletonPath,
 } from '../path-utils';
-import { getTreeNodeAtPath, treeEntriesToTreeNodes, TreeEntry, TreeNode } from '../trees';
+import { getTreeNodeAtPath, treeEntriesToTreeNodes, TreeEntry, TreeNode, treeSha } from '../trees';
 import { DataState, LOADING, mergeDataStates, useData } from '../useData';
 import { getTreeNodeForItem, MaybePromise } from '../utils';
 import LRU from 'lru-cache';
 import { isDefined } from 'emery';
 
-export function useAppShellData(props: { currentBranch: string; config: Config }) {
+export function fetchLocalTree(sha: string) {
+  if (treeCache.has(sha)) {
+    return treeCache.get(sha)!;
+  }
+  const promise = fetch('/api/keystatic/tree', { headers: { 'no-cors': '1' } })
+    .then(x => x.json())
+    .then(async (entries: TreeEntry[]) => hydrateTreeCacheWithEntries(entries));
+  treeCache.set(sha, promise);
+  return promise;
+}
+
+export function useSetTreeSha() {
+  return useContext(SetTreeShaContext);
+}
+
+export const SetTreeShaContext = createContext<(sha: string) => void>(() => {
+  throw new Error('SetTreeShaContext not set');
+});
+
+export function LocalAppShellProvider(props: { config: LocalConfig; children: ReactNode }) {
+  const [currentTreeSha, setCurrentTreeSha] = useState<string>('initial');
+
+  const tree = useData(useCallback(() => fetchLocalTree(currentTreeSha), [currentTreeSha]));
+
+  const allTreeData = useMemo(
+    () => ({
+      default: tree,
+      current: tree,
+      merged: mergeDataStates({ default: tree, current: tree }),
+    }),
+    [tree]
+  );
+  const changedData = useMemo(() => {
+    if (allTreeData.merged.kind !== 'loaded') {
+      return {
+        collections: new Map<
+          string,
+          { removed: Set<string>; added: Set<string>; changed: Set<string>; totalCount: number }
+        >(),
+        singletons: new Set<string>(),
+      };
+    }
+    return getChangedData(props.config, allTreeData.merged.data);
+  }, [allTreeData, props.config]);
+
+  return (
+    <SetTreeShaContext.Provider value={setCurrentTreeSha}>
+      <ChangedContext.Provider value={changedData}>
+        <TreeContext.Provider value={allTreeData}>{props.children}</TreeContext.Provider>
+      </ChangedContext.Provider>
+    </SetTreeShaContext.Provider>
+  );
+}
+
+export function GitHubAppShellProvider(props: {
+  currentBranch: string;
+  config: GitHubConfig;
+  children: ReactNode;
+}) {
   const router = useRouter();
   const [{ data, error }] = useQuery({
     query: AppShellQuery,
     variables: {
-      branch: props.currentBranch,
-      name: props.config.repo.name,
-      owner: props.config.repo.owner,
+      name: props.config.storage.repo.name,
+      owner: props.config.storage.repo.owner,
     },
   });
   const defaultBranchRef = data?.repository?.refs?.nodes?.find(
@@ -38,8 +103,8 @@ export function useAppShellData(props: { currentBranch: string; config: Config }
   const currentBranchTreeSha = currentBranchRef?.target.tree.oid ?? null;
   const baseCommit = currentBranchRef?.target?.oid ?? null;
 
-  const defaultBranchTree = useTreeData(defaultBranchTreeSha, props.config.repo);
-  const currentBranchTree = useTreeData(currentBranchTreeSha, props.config.repo);
+  const defaultBranchTree = useGitHubTreeData(defaultBranchTreeSha, props.config.storage.repo);
+  const currentBranchTree = useGitHubTreeData(currentBranchTreeSha, props.config.storage.repo);
 
   const allTreeData = useMemo(
     () => ({
@@ -59,56 +124,7 @@ export function useAppShellData(props: { currentBranch: string; config: Config }
         singletons: new Set<string>(),
       };
     }
-    const trees = allTreeData.merged.data;
-
-    return {
-      collections: new Map(
-        Object.keys(props.config.collections ?? {}).map(collection => {
-          const collectionPath = getCollectionPath(props.config, collection);
-          const current = new Map(getTreeNodeAtPath(trees.current.tree, collectionPath)?.children);
-          const defaultBranch = new Map(
-            getTreeNodeAtPath(trees.default.tree, collectionPath)?.children
-          );
-
-          const formatInfo = getCollectionFormat(props.config, collection);
-          const extension = getDataFileExtension(formatInfo);
-          for (const map of [current, defaultBranch]) {
-            for (const [key, entry] of map) {
-              const node = getTreeNodeForItem(props.config, collection, entry);
-              if (!node || !node.children?.has(`index${extension}`)) {
-                map.delete(key);
-              } else {
-                map.set(key, node);
-              }
-            }
-          }
-
-          const changed = new Set<string>();
-          const added = new Set<string>();
-          for (const [key, entry] of current) {
-            const defaultBranchEntry = defaultBranch.get(key);
-            if (defaultBranchEntry === undefined) {
-              added.add(key);
-              continue;
-            }
-            if (entry.entry.sha !== defaultBranchEntry.entry.sha) {
-              changed.add(key);
-            }
-          }
-          const removed = new Set([...defaultBranch.keys()].filter(key => !current.has(key)));
-          return [collection, { removed, added, changed, totalCount: current.size }];
-        })
-      ),
-      singletons: new Set(
-        Object.keys(props.config.singletons ?? {}).filter(singleton => {
-          const singletonPath = getSingletonPath(props.config, singleton);
-          return (
-            getTreeNodeAtPath(trees.current.tree, singletonPath)?.entry.sha !==
-            getTreeNodeAtPath(trees.default.tree, singletonPath)?.entry.sha
-          );
-        })
-      ),
-    };
+    return getChangedData(props.config, allTreeData.merged.data);
   }, [allTreeData, props.config]);
 
   useEffect(() => {
@@ -136,6 +152,9 @@ export function useAppShellData(props: { currentBranch: string; config: Config }
       repositoryId: data?.repository?.id ?? '',
       allBranches: data?.repository?.refs?.nodes?.map(x => x?.name).filter(isDefined) ?? [],
       hasPullRequests: !!currentBranchRef?.associatedPullRequests.totalCount,
+      branchNameToId: new Map(
+        data?.repository?.refs?.nodes?.filter(isDefined).map(x => [x.name, x.id])
+      ),
     }),
     [
       data?.repository?.defaultBranchRef?.name,
@@ -146,19 +165,20 @@ export function useAppShellData(props: { currentBranch: string; config: Config }
       currentBranchRef?.associatedPullRequests.totalCount,
     ]
   );
-  return {
-    error,
-    providers: (children: ReactNode) => (
+  return (
+    <AppShellErrorContext.Provider value={error}>
       <BranchInfoContext.Provider value={branchInfo}>
         <BaseInfoContext.Provider value={baseInfo}>
           <ChangedContext.Provider value={changedData}>
-            <TreeContext.Provider value={allTreeData}>{children}</TreeContext.Provider>
+            <TreeContext.Provider value={allTreeData}>{props.children}</TreeContext.Provider>
           </ChangedContext.Provider>
         </BaseInfoContext.Provider>
       </BranchInfoContext.Provider>
-    ),
-  };
+    </AppShellErrorContext.Provider>
+  );
 }
+
+export const AppShellErrorContext = createContext<Error | undefined>(undefined);
 
 const BaseInfoContext = createContext({ baseCommit: '', repositoryId: '' });
 
@@ -212,7 +232,7 @@ export function useRepositoryId() {
   return useContext(BaseInfoContext).repositoryId;
 }
 
-const AppShellQuery = gql`
+export const AppShellQuery = gql`
   query AppShell($name: String!, $owner: String!) {
     repository(owner: $owner, name: $name) {
       id
@@ -253,16 +273,17 @@ const treeCache = new LRU<
   max: 40,
 });
 
-export function hydrateTreeCacheWithEntries(sha: string, entries: TreeEntry[]) {
+export async function hydrateTreeCacheWithEntries(entries: TreeEntry[]) {
   const data = {
     entries: new Map(entries.map(entry => [entry.path, entry])),
     tree: treeEntriesToTreeNodes(entries),
   };
+  const sha = await treeSha(data.tree);
   treeCache.set(sha, data);
   return data;
 }
 
-export function fetchTreeData(sha: string, repo: { owner: string; name: string }) {
+export function fetchGitHubTreeData(sha: string, repo: { owner: string; name: string }) {
   const cached = treeCache.get(sha);
   if (cached) return cached;
   const promise = githubRequest('GET /repos/{owner}/{repo}/git/trees/{tree_sha}', {
@@ -271,22 +292,20 @@ export function fetchTreeData(sha: string, repo: { owner: string; name: string }
     tree_sha: sha,
     recursive: '1',
   }).then(res =>
-    hydrateTreeCacheWithEntries(
-      sha,
-      res.data.tree.map(({ url, ...rest }) => rest as TreeEntry)
-    )
+    hydrateTreeCacheWithEntries(res.data.tree.map(({ url, ...rest }) => rest as TreeEntry))
   );
   treeCache.set(sha, promise);
   return promise;
 }
 
-function useTreeData(sha: string | null, repo: { owner: string; name: string }) {
-  return useData(useCallback(() => (sha ? fetchTreeData(sha, repo) : LOADING), [sha, repo]));
+function useGitHubTreeData(sha: string | null, repo: { owner: string; name: string }) {
+  return useData(useCallback(() => (sha ? fetchGitHubTreeData(sha, repo) : LOADING), [sha, repo]));
 }
 
 export const BranchInfoContext = createContext<{
   currentBranch: string;
   allBranches: string[];
+  branchNameToId: Map<string, string>;
   defaultBranch: string;
   hasPullRequests: boolean;
 }>({
@@ -294,4 +313,56 @@ export const BranchInfoContext = createContext<{
   allBranches: [],
   defaultBranch: '',
   hasPullRequests: false,
+  branchNameToId: new Map(),
 });
+
+function getChangedData(config: Config, trees: { current: TreeData; default: TreeData }) {
+  return {
+    collections: new Map(
+      Object.keys(config.collections ?? {}).map(collection => {
+        const collectionPath = getCollectionPath(config, collection);
+        const current = new Map(getTreeNodeAtPath(trees.current.tree, collectionPath)?.children);
+        const defaultBranch = new Map(
+          getTreeNodeAtPath(trees.default.tree, collectionPath)?.children
+        );
+
+        const formatInfo = getCollectionFormat(config, collection);
+        const extension = getDataFileExtension(formatInfo);
+        for (const map of [current, defaultBranch]) {
+          for (const [key, entry] of map) {
+            const node = getTreeNodeForItem(config, collection, entry);
+            if (!node || !node.children?.has(`index${extension}`)) {
+              map.delete(key);
+            } else {
+              map.set(key, node);
+            }
+          }
+        }
+
+        const changed = new Set<string>();
+        const added = new Set<string>();
+        for (const [key, entry] of current) {
+          const defaultBranchEntry = defaultBranch.get(key);
+          if (defaultBranchEntry === undefined) {
+            added.add(key);
+            continue;
+          }
+          if (entry.entry.sha !== defaultBranchEntry.entry.sha) {
+            changed.add(key);
+          }
+        }
+        const removed = new Set([...defaultBranch.keys()].filter(key => !current.has(key)));
+        return [collection, { removed, added, changed, totalCount: current.size }];
+      })
+    ),
+    singletons: new Set(
+      Object.keys(config.singletons ?? {}).filter(singleton => {
+        const singletonPath = getSingletonPath(config, singleton);
+        return (
+          getTreeNodeAtPath(trees.current.tree, singletonPath)?.entry.sha !==
+          getTreeNodeAtPath(trees.default.tree, singletonPath)?.entry.sha
+        );
+      })
+    ),
+  };
+}